split out header components
