+++ easy-ui-react/src/FocusedProductLayout/WizardContent.module.scss
@@ -10,7 +10,7 @@
 
 .stepperContainer {
   flex: 0 0 auto;
-  padding-bottom: design-token("space.1.5");
+  padding-bottom: design-token("space.1");
   border-bottom: design-token("shape.border_width.1") solid
     design-token("color.gray.200");
 }
